project-inventory/src/main/java/com/enigma/arinda/model/transactions/TransactionResponse.java
package com.enigma.arinda.model.transactions;

import com.enigma.arinda.entities.User;
import com.enigma.arinda.model.items.ItemResponse;
import com.enigma.arinda.model.users.UserModel;

public class TransactionResponse {
    private Integer id;
    private ItemResponse itemResponse;
    private UserModel userModel;

    private Integer quantity;
    private Integer totalPrice;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public ItemResponse getItemResponse() {
        return itemResponse;
    }

    public void setItemResponse(ItemResponse itemResponse) {
        this.itemResponse = itemResponse;
    }

    public Integer getQuantity() {
        return quantity;
    }

    public void setQuantity(Integer quantity) {
        this.quantity = quantity;
    }

    public Integer getTotalPrice() {
        return totalPrice;
    }

    public void setTotalPrice(Integer totalPrice) {
        this.totalPrice = totalPrice;
    }

    public UserModel getUserModel() {
        return userModel;
    }

    public void setUserModel(UserModel userModel) {
        this.userModel = userModel;
    }
}
